requirements.txt
numpy==2.2.6
scipy==1.16.3
Pillow==12.1.0
torch==2.8.0
torchvision==0.23.0
psutil==7.1.3
loguru==0.7.3
imageio==2.37.2
tqdm==4.67.1
easydict==1.13
opencv-python-headless==4.12.0.88
trimesh==4.11.0
transformers==4.57.3
tensorboard==2.20.0
pandas==2.3.3
lpips==0.1.4
zstandard==0.25.0
fastapi==0.128.0
uvicorn==0.40.0
kornia==0.8.2
timm==1.0.24
python-multipart==0.0.21
packaging==25.0
einops==0.8.1
# Build: pybind11 for mesh_inpaint_processor compile
pybind11>=2.10.0

diffusers==0.36.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic-tensor==0.2.0
safetensors==0.7.0
huggingface-hub==0.36.0
accelerate==1.12.0
peft==0.18.1
# Hunyuan shape (marching cubes, flow matching, config)
scikit-image>=0.24.0
torchdiffeq>=0.2.3
pyyaml>=6.0

# Hunyuan paint (PBR GLB, texture, remesh, UV, renderer mesh I/O)
pygltflib>=1.16.0
omegaconf>=2.3.0
realesrgan>=0.3.0
basicsr>=1.4.2
pymeshlab>=2022.2
xatlas>=0.0.9
pytorch-lightning>=2.0.0
fast-simplification>=0.1.0